backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.0
django-cors-headers>=4.3
django-extensions>=3.2
celery>=5.3
redis>=5.0
psycopg[binary]>=3.1
python-docx>=1.1
WeasyPrint>=62.0
qrcode[pil]>=7.4
Pillow>=10.0
minio>=7.2
gunicorn>=22.0
openpyxl>=3.1
python-barcode>=0.15
reportlab>=4.0
